requirements.txt
Django==2.2.24
psycopg2-binary==2.8.6
djangorestframework==3.11.2
django-rest-swagger==2.2.0
drf_yasg
pandas
xlrd==1.2.0
django-import-export==2.0.2
django-admin-select2
pymongo==3.9.0
PyJWT==1.7.1
wget==3.2
sklearn
celery
redis
numpy==1.17.2
django_extensions
gunicorn==20.0.4
openpyxl==3.0.3
#Pillow==2.2.1
statsmodels
django-braces==1.14.0
#django-weasyprint
#django-wkhtmltopdf
xhtml2pdf

#TODO Cesar quitar django_extensions
